[FUNCTION ] New way to fusion gpsyaw.

+++ src/modules/ekf2/ekf2_main.cpp
@@ -1024,7 +1024,7 @@ void Ekf2::run()
 			if(orb_copy(ORB_ID(vehicle_gps_yaw), _vehicle_gps_yaw_sub, &vehicle_gps_yaw) == PX4_OK) {
 				if (vehicle_gps_yaw.fix_type == 6) {
 
-					_ekf.setGpsYawData(vehicle_gps_yaw.timestamp, wrap_pi(vehicle_gps_yaw.yaw * M_PI / 180.0f));
+					_ekf.setGpsYawData(vehicle_gps_yaw.timestamp, wrap_pi((float)(vehicle_gps_yaw.yaw * (float)M_PI / 180.0f)));
 
 				}
 			}
